pyproject.toml
[project]
name = "cosmosdb-python-101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.21.0",
    "python-dotenv>=1.1.0",
]
